Fix fork choice logic - move processing out of tx.commit()
