+++ eth-reference-tests/build.gradle
@@ -5,6 +5,7 @@ dependencies {
   referenceTestImplementation project(':data')
   referenceTestImplementation project(':ethereum:core')
   referenceTestImplementation project(':ethereum:datastructures')
+  referenceTestImplementation project(':ethereum:statetransition')
   referenceTestImplementation project(':eth-tests')
   referenceTestImplementation project(':protoarray')
   referenceTestImplementation project(':ssz')
