+++ eth-reference-tests/src/referenceTest/java/tech/pegasys/teku/reference/phase0/fork_choice/implementatations/ForkChoiceProcessor.java
@@ -1,0 +1,35 @@
+/*
+ * Copyright 2020 ConsenSys AG.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
+ * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations under the License.
+ */
+
+package tech.pegasys.teku.reference.phase0.fork_choice.implementatations;
+
+import org.apache.tuweni.bytes.Bytes32;
+import tech.pegasys.teku.core.StateTransition;
+import tech.pegasys.teku.datastructures.blocks.SignedBeaconBlock;
+import tech.pegasys.teku.datastructures.operations.Attestation;
+import tech.pegasys.teku.storage.client.RecentChainData;
+
+public abstract class ForkChoiceProcessor {
+  protected final RecentChainData chainData;
+  protected final StateTransition st = new StateTransition();
+
+  protected ForkChoiceProcessor(final RecentChainData chainData) {
+    this.chainData = chainData;
+  }
+
+  public abstract boolean processBlock(SignedBeaconBlock block);
+
+  public abstract boolean processAttestation(Attestation attestation);
+
+  public abstract Bytes32 processHead();
+}
